Fixed productCards axios
