--- .env
@@ -1,5 +1,0 @@
-PORT=9000
-
-API_URL='https://app-hrsei-api.herokuapp.com/api/fec2/hr-rfe/'
-
-TOKEN='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'
