+++ client/src/components/relatedProducts/productCards.jsx
@@ -1,17 +1,42 @@
-import React from 'react';
+import React, {useEffect} from 'react';
 import axios from 'axios';
 import RelatedList from './relatedList.jsx';
 
-const ProductCard = ({styles, photo, product}) => {
+const ProductCard = ({styles, photo, id}) => {
+  const [product, setProduct] = React.useState(null);
+
+  useEffect(() => {
+    const page = 1;
+    const fetcher = (page) => {
+      if (product) {
+        return;
+      }
+      axios.get(`http://localhost:9000/products?page=${page}&count=10`)
+        .then(response => {
+          console.log('respondes', response.data)
+          for (var each of response.data) {
+            console.log('ththeeid', typeof id)
+            if (each.id === id) {
+              console.log('got one!', each.id)
+              setProduct(each);
+              return;
+            }
+          }
+          page++
+          fetcher(page);
+        })
+        .catch(error =>
+          console.log('An error fetching from server:', error));
+    };
+    fetcher(page);
+    console.log('new useeffect', product)
+  }, [id]);
 
   if (!styles || !photo || !product) {
     return null;
   }
 
-  const name = styles.name;
   const price = styles.original_price;
-
-  console.log('me product', product)
 
 
   return (
@@ -20,8 +45,8 @@ const ProductCard = ({styles, photo, product}) => {
         <img className='cardImage' src={photo} />
       </div>
       <div className='container'>
-        <h2>CATEGORY</h2>
-        <h3><b>{name}</b></h3>
+        <h2>{product.category}</h2>
+        <h3><b>{product.name}</b></h3>
         <p>${price}</p>
       Rating
       </div>
